Renamed file from .hpp to .h
